Made client non-static for potential future integration of byngosinc and lockout live servers, added chat history and chat message event interfaces
