+++ BingoSync/Clients/ChatMessage.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace BingoSync.Clients
+{
+    public class ChatMessage : EventArgs
+    {
+        public string Text { get; set; }
+        public string Sender { get; set; }
+        public DateTime TimeStamp { get; set; } = DateTime.Now;
+    }
+}
